03.1 de junio

+++ src/main/java/org/iesbelen/exception/InstrumentoNotFoundException.java
@@ -1,7 +1,7 @@
 package org.iesbelen.exception;
 
 public class InstrumentoNotFoundException extends RuntimeException{
-    public InstrumentoNotFoundException(Long id) {
+    public InstrumentoNotFoundException(int id) {
         super("Not found Instrumento with id: " + id);
     }
 }
